# ert-leader-ui/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8912",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/admin/question/{questionId}": {
      "put": {
        "tags": [
          "题目管理"
        ],
        "summary": "更新题目",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "题目管理"
        ],
        "summary": "删除题目",
        "description": "级联删除关联的选项和答案数据",
        "operationId": "delete",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "认证与用户模块"
        ],
        "summary": "用户注册",
        "description": "新用户注册，密码需满足8位以上含字母数字",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证与用户模块"
        ],
        "summary": "用户登录",
        "description": "登录成功后返回JWT token，有效期待定",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoginVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/question": {
      "post": {
        "tags": [
          "题目管理"
        ],
        "summary": "创建题目",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultCreateQuestionVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/question/page": {
      "get": {
        "tags": [
          "题目管理"
        ],
        "summary": "分页查询题目",
        "operationId": "page_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "subDimCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "type": "object",
            "description": "数据"
          }
        }
      },
...
      "CreateQuestionDTO": {
        "type": "object",
        "properties": {
          "subDimCode": {
            "type": "string",
            "description": "子维度编码"
          },
          "content": {
            "type": "string",
            "description": "题干"
          },
          "seq": {
            "type": "integer",
            "description": "排序",
            "format": "int32"
          }
        }
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string",
            "description": "手机号（唯一）"
          },
          "email": {
            "type": "string",
            "description": "邮箱（唯一）"
          },
          "password": {
            "type": "string",
            "description": "密码（bcrypt 加密）"
          },
          "name": {
            "type": "string",
            "description": "真实姓名"
          },
          "gender": {
            "type": "integer",
            "description": "性别：0 未知，1 男，2 女",
            "format": "int32"
          },
          "birthDate": {
            "type": "string",
            "description": "出生日期",
            "format": "date"
          }
        },
        "description": "用户信息DTO"
      },
      "ResultUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "$ref": "#/components/schemas/UserVO"
          }
        }
      },
      "UserVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户主键",
            "format": "int64"
          },
          "mobile": {
            "type": "string",
            "description": "手机号（唯一）"
          },
          "email": {
            "type": "string",
            "description": "邮箱（唯一）"
          },
          "name": {
            "type": "string",
            "description": "真实姓名"
          },
          "role": {
            "type": "string",
            "description": "角色：USER/INST_ADMIN/SUPER_ADMIN"
          }
        },
        "description": "数据"
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "账号"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        }
      },
      "LoginVO": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token"
          },
          "expiresIn": {
            "type": "string",
            "description": "token过期时间"
          },
          "userInfo": {
            "$ref": "#/components/schemas/UserInfo"
          }
        },
        "description": "数据"
      },
      "ResultLoginVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "$ref": "#/components/schemas/LoginVO"
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "用户名"
          },
          "role": {
            "type": "string",
            "description": "用户角色"
          }
        },
        "description": "用户信息"
      },
...
      "CreateQuestionVO": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "description": "问题ID",
            "format": "int64"
          }
        },
        "description": "数据"
      },
      "ResultCreateQuestionVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "$ref": "#/components/schemas/CreateQuestionVO"
          }
        }
      },
...
      "PageResult": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "总记录数",
            "format": "int64",
            "refType": null
          },
          "records": {
            "type": "array",
            "description": "当前页数据",
            "items": {
              "type": "object",
              "description": "当前页数据"
            },
            "refType": "object"
          }
        },
        "description": "数据"
      },
      "ResultPageResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "$ref": "#/components/schemas/PageResult"
          }
        }
      },
...
    }
  },
...
}